requirements.txt
streamlit
pandas
streamlit-lottie
plotly
pyarrow
scikit-surprise